requirements.txt
numpy==1.19.4
opencv-python==4.4.0.46
torch==1.7.0
torchvision==0.8.1
pillow==8.0.1
pdf2image==1.14.0
pyyaml==5.3.1
